重构服务层，更新订单和员工相关方法，修复DTO字段类型，移除不必要的导入

+++ BackendService/src/main/java/com/example/dto/OrderDTO.java
@@ -1,6 +1,5 @@
 package com.example.dto;
 
-import com.example.model.customer.Customer;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderDTO {
-    private int id;
+    private Long id;
     @NotBlank
     private LocalDate createDate;
     @NotBlank
